Move DBConnector to database package
